% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2DF.Long.R
\name{Dist2DF.Long}
\alias{Dist2DF.Long}
\title{`dist` to `data.frame` long format conversion}
\usage{
Dist2DF.Long(inDist, Convert100 = FALSE)
}
\arguments{
\item{inDist}{A matrix of class dist, as the output of `ape::dist.dna`.}

\item{Convert100}{If TRUE converts proportion distance (or divergence) input to percentages (`FASLE` is the default).}
}
\value{
The distance data in class 'data.frame' long format
}
\description{
A function to change the data class from `dist` to `data.frame` in long format.
}
\examples{
library(ape)

test.dist <- dist.dna(test.DNA)

Dist.df <- Dist2DF.Long(test.dist)

Dist.df.100 <- Dist2DF.Long(test.dist, Convert100 = TRUE)

Dist.df.100x <- Dist2DF.Long(test.dist*100) # equivalent to `Convert100 = TRUE

}
