% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.Dist.df.R
\name{DNA.Dist.df}
\alias{DNA.Dist.df}
\title{Single locus divergence data.frame}
\usage{
DNA.Dist.df(
  inDNA,
  Dist.type = "percent",
  Out.Format = "long",
  Model = "K80",
  GAMMA = FALSE,
  PW.deletion = TRUE,
  Var = FALSE,
  by.group = NULL,
  group.summary = FALSE,
  within.group = FALSE
)
}
\arguments{
\item{inDNA}{the input aligned sequences in `.fasta` format.}

\item{Dist.type}{specifies if the distance output should be in proportions (`"prop"`), percentages (`"percentage"`), or number of nucleotides that are different (`"nucleotide"`).}

\item{Model}{specifies the evolutionary model to use. Same as in `ape::dist.dna`. Options are: `"RAW"`, `"JC69"`, `"K80"` (default), `"F81"`, `"K81"`, `"F84"`, `"T92"`, `"TN93"`, `"GG95"`, `"LOGDET"`, `"BH87"`, `"PARALIN"`, `"N"`, `"TS"`, `"TV"`, `"INDEL"`, and `"INDELBLOCK"`. Note: if `Model` is `"N"` and `Dist.type` is set for `"prop"` or `"percentage"`, then an error will be return.}

\item{GAMMA}{same as `gamma` in `ape::dist.dna()`.}

\item{PW.deletion}{specifies if Pairwise deletion should be considered (`TURE`, default) or not (`FALSE`).}

\item{Var}{if `TRUE` output is the calculated variation (`FALSE` is default)}

\item{by.group}{a vector to group sequences on different categories (e.g., populations, species, lineages, etc.). The vector should follow the same order as the sequence alignment, by applying `vegan::meandist()`. The vector can be the `.csv` output of species delimitation analyses such as `ASAP` and `ABGD`.}

\item{`group.summary`}{applies `vegan::meandist()` to formulate a data.frame with within groups mean distance, between groups mean distance and overall distance. A vector with discreet groupings should be imported in `by.group`.}

\item{`within.group`}{applies `vegan::meandist()` to create a `data.frame` with each within group mean divergence, and number of samples ("N") by group. A vector with discreet groupings should be imported in `by.group`. Groupls with N = 1 will return `NA`'s.}
}
\value{
A dataframe with the mean distance between groups or within group distances.
         Note: `DNA.Dist.df` should not be used as a species delimitation analysis as it does not explicitly tests a hypothesis.
}
\description{
A function that applies `ape::dist.dna()` but changes `"dist"` class output from `ape::dist.dna()` (or similar data) to `data.frame`.
}
\examples{
library("ape")
#UNder construction!

}
