% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.GroupDist.SE.R
\name{DNA.GroupDist.SE}
\alias{DNA.GroupDist.SE}
\title{Divergence Standard Error}
\usage{
DNA.GroupDist.SE(
  inDNA,
  Dist.type = "percent",
  Out.Format = "long",
  Model = "K80",
  GAMMA = FALSE,
  PW.deletion = TRUE,
  by.group,
  within.group = FALSE
)
}
\arguments{
\item{inDNA}{the imput aligned sequences in `.fasta` format.}

\item{Dist.type}{specifies if the distance output should be in proportions (`"prop"`), percentages (`"percentage"`), or number of nucleotides that are different (`"nucleotide"`).}

\item{Model}{specifies the evolutionary model to use. Same as in `ape::dist.dna`. Options are: `"RAW"`, `"JC69"`, `"K80"` (default), `"F81"`, `"K81"`, `"F84"`, `"T92"`, `"TN93"`, `"GG95"`, `"LOGDET"`, `"BH87"`, `"PARALIN"`, `"N"`, `"TS"`, `"TV"`, `"INDEL"`, and `"INDELBLOCK"`. Note: if `Model` is `"N"` and `Dist.type` is set for `"prop"` or `"percentage"`, then an error will be return.}

\item{GAMMA}{same as `gamma` in `ape::dist.dna()`.}

\item{PW.deletion}{specifies if Pairwise deletion should be considered (`TURE`, default) or not (`FALSE`).}

\item{by.group}{a vector to group sequences on different categories (e.g., populations, species, lineages, etc.). The vector should follow the same order as the sequence alignment. The vector can be the `.csv` output of species delimitation analyses such as `ASAP` and `ABGD`.}

\item{within.group}{creates a `data.frame` with each within group standard error, and number of saples ("N") by group. A vector with discreet groupings should be imported in `by.group`. Groups with N = 1 will return `NA`'s.}
}
\value{
A data.frame with the standard errors of the mean distance between or within groups distances.
}
\description{
Calculates standard error of specified between and within group genetic divergences.
}
