% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE}
\alias{SE}
\title{Standard Error}
\usage{
SE(x)
}
\arguments{
\item{x}{a numeric vector. `NA` will be automatically ignored.}
}
\value{
Standard error calculation
}
\description{
Calculates standard error on a vector
}
\examples{
a <- c(1:10)
SE(a)

a <- c(1:10, NA, NA)
SE(a)

a <- c(NA, NA, NA)
SE(a)

}
