% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.DNA.R
\docType{data}
\name{test.DNA}
\alias{test.DNA}
\title{test.DNA}
\format{
A list of nine randomly simulated DNA dataset of three specimens for three "species", composed of 50 bp each.

    Sequence names are divided by the specimen's ID as IndA, IndB, ... and species ID as Sp1, Sp2 ... separated by an underscore.
}
\usage{
test.DNA
}
\description{
Binary list of simulated DNA sequences to test funtions
}
\keyword{datasets}
