% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2DF.Groups.R
\name{Dist2DF.Groups}
\alias{Dist2DF.Groups}
\title{Mean Distances by Group}
\usage{
Dist2DF.Groups(inDist, Convert100 = FALSE, get.group = NULL, n.groups = NULL)
}
\arguments{
\item{Convert100}{Converts proportions to percentages.}

\item{get.group}{Provides discreet grouping using maximum divergence threshold between sequences.}

\item{n.groups}{Groups low-diverging sequences to a specified number of groups, regardless their within group divergences.}

\item{inDust}{A `"dist"` object matrix, output of `ape::dist.dna` or alike matrix}
}
\value{
A data.frame with the mean distance between groups group distances from a distance matrix produced by `ape::dist.dna`
     Note: `Dist2DF.Groups` should not be used as a species delimitation analysis as it does not explicitly tests a hypothesis.
}
\description{
A function that applies `hclust` and `cutree` to a matrix of class `"dist"` to cluster samples by a  maximum distance threshold or to a set number of groups.
}
