% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2DF.Long.R
\name{Dist2DF.Long}
\alias{Dist2DF.Long}
\title{`dist` to `data.frame` long format conversion}
\usage{
Dist2DF.Long(inDist, Convert100 = FALSE)
}
\arguments{
\item{inDist}{A matrix of class dist, as the output of `ape::dist.dna`.}

\item{Convert100}{If TRUE converts proportion distance (or divergence) input to percentages (`FASLE` is the default).}
}
\value{
The distance data in class 'data.frame' long format
}
\description{
A function to change the data class from `dist` to `data.frame` in long format.
}
\examples{
library("ape")
COI.data <- read.FASTA(file="COI_aln.fasta")
Dist <- dist.dna(COI.data, model = "K80")

#To obtain just a pairwise long format table of proportional distance:
long_dist1 <- Dist2DF.Long(Dist)
head(long_dist1)

#To obtain just a pairwise long format table of distances in percentages:
Dist <- dist.dna(COI.data, model = "K80", variance = TRUE)
long_dist2 <- Dist2DF.Long(Dist, Convert100 = TRUE)
head(long_dist2)

}
