% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2DF.Wide.R
\name{Dist2DF.Wide}
\alias{Dist2DF.Wide}
\title{'dist' to 'data.frame' wide format conversion}
\usage{
Dist2DF.Wide(inDist, Convert100 = FALSE)
}
\arguments{
\item{inDist}{A matrix of class `dist`, as the output of `ape::dist.dna`.}

\item{Convert100}{If `TRUE` converts proportion distance (or divergence) input to percentages (`FASLE` is the default).}
}
\value{
The distance data in class 'data.frame' wide format.
}
\description{
A function to change the data class from dist to data.frame in wide format.
}
\examples{
library("ape")
COI.data <- read.FASTA(file="COI.fasta")
Dist <- dist.dna(COI.data, model = "K80")

# Wide matrix-like dataframe of proportional distances
DistW1 <- Dist2DF.Wide(Dist)
head(DistW1)

# Wide matrix-like dataframe of distances in percentages
DistW2 <- Dist2DF.Wide(Dist, Convert100 = TRUE)
head(DistW2)

}
